Bill S1299 version original - Leroy Comrie
BILL NUMBER: S1299 SPONSOR: COMRIE TITLE OF BILL: An act to amend the tax law, in relation to establishing a personal income tax credit for tolls paid by a taxpayer in the course of commut- ing to and from his or her place of employment PURPOSE: To provide relief from high toll expenses by creating an income tax credit for individuals that must pay tolls as part of their daily commute. SUMMARY OF PROVISIONS: Section 1.