Bill S1375 version original - Simcha Felder
BILL NUMBER: S1375 SPONSOR: FELDER TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to requiring all persons in a city having a population of one million or more operat- ing a bicycle, electric bicycle or electric scooter to wear a helmet PURPOSE: To require every person operating a bicycle, E-Bike or E-Scooter in New York City to wear a helmet. SUMMARY OF PROVISIONS: Section one adds a new section 114-f to the vehicle and traffic law that defines an electric bicycle.