Bill S1296 version original - James Sanders Jr.
BILL NUMBER: S1296 SPONSOR: SANDERS TITLE OF BILL: An act to amend the elder law, in relation to informal caregiver train- ing PURPOSE OR GENERAL IDEA OF BILL: To encourage the development of training and supports for older adults who are caregivers for minors. SUMMARY OF SPECIFIC PROVISIONS: This bill would broaden the informal caregiver training program to include training and resources for older adults who are caregivers for minors.