Bill S1359 version original - Timothy M. Kennedy
BILL NUMBER: S1359 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the executive law, in relation to establishing the office of racial equity and social justice PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to create an Office of Racial Equity and Social Justice. SUMMARY OF PROVISIONS: Section one adds a new article 15-D to the executive law which provides definitions of terms and outlines the powers, functions, and duties of the Office of Racial Equity and Social Justice.