Bill S1391 version original - James Tedisco
BILL NUMBER: S1391 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the penal law, in relation to elevating the severity of the felony offense of killing a police work dog or police work horse and establishing the class E felony offense of injuring a police work dog or police work horse PURPOSE: The purpose of this bill is to increase penalties for the injury or killing of a police work dog or police work horse. SUMMARY OF PROVISIONS: Section 1: Amends Subdivision section 195.