Bill S1431 version original - Jamaal Bailey
BILL NUMBER: S1431 SPONSOR: BAILEY TITLE OF BILL: An act to amend chapter 137 of the laws of 1817 relating to slaves and servants, in relation to acknowledging the tragedy of slavery in New York state PURPOSE: This bill acknowledges the tragedy of slavery in New York state and offers a formal apology from the State of New York for its role in perpetuating slavery. SUMMARY OF PROVISIONS: Section 1.