Bill S1349 version original - Gustavo Rivera
BILL NUMBER: S1349 SPONSOR: RIVERA TITLE OF BILL: An act to amend the public health law and the social services law, in relation to authorizing residents of assisted living programs to receive hospice services PURPOSE: This is a chapter amendment that makes a technical change to provisions of L.2022, c.667. SUMMARY OF PROVISIONS: This legislation makes a technical change to the underlying bill by requiring the State to apply for federal matching funds for these hospice services.