Bill S1275 version original - Robert Ortt
BILL NUMBER: S1275 SPONSOR: ORTT TITLE OF BILL: An act to amend the public authorities law, in relation to adding two trustees to the power authority to represent Niagara county and St. Lawrence county PURPOSE: Relates to adding two trustees to the power authority to represent Niagara county and St. Lawrence county. SUMMARY OF SPECIFIC PROVISIONS: Section 1. - Increases the number of trustees to the Power Authority from seven to nine.