Bill S1438 version original - Joseph A. Griffo
BILL NUMBER: S1438 SPONSOR: GRIFFO TITLE OF BILL: CONCURRENT RESOLUTION OF THE SENATE AND ASSEMBLY proposing an amendment to article 4 of the constitution, relating to the lieutenant-governor acting as governor PURPOSE: This bill would remove the constitutional provision that the lieuten- ant-governor shall act as governor whenever the governor is absent from the state. SUMMARY OF PROVISIONS: § 1 - Amends the third undesignated paragraph of § 5 of Article 4 of the Constitution.