Bill S1252 version original - Rachel May
BILL NUMBER: S1252 SPONSOR: MAY TITLE OF BILL: An act to amend the education law, in relation to payments made by school districts to charter schools PURPOSE OR GENERAL IDEA OF BILL: This bill requires that reimbursements to school districts for payments to charter schools mirror those payments. SUMMARY OF SPECIFIC PROVISIONS: Section 1 - Paragraph (b) of subdivision 1 of section 2856 of the educa- tion law, as amended by section 4 of part YYY of chapter 59 of the laws of 2017, is amended.