Bill S1393 version original - James Tedisco
BILL NUMBER: S1393 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to requiring that license plates be inspected during safety inspections and waiving fees for the issuance of certain license plates PURPOSE: To establish a process to inspect New York license plates for legibili- ty, and to waive fees for certain license plates. SUMMARY OF PROVISIONS: § 1.