Bill S1399 version original - John Liu
BILL NUMBER: S1399 SPONSOR: LIU TITLE OF BILL: An act to require any public school located in a city with a population of one million or more to provide a remote learning option when such public school is located in a county where community transmission of COVID-19 is at a substantial or high level SUMMARY OF PROVISIONS: Section 1. Requires that a public school located in a county with substantial or high levels of community transmission of COVID-19 provide a remote learning option.