Bill S1295A version original - James Sanders Jr.
BILL NUMBER: S1295 SPONSOR: SANDERS TITLE OF BILL: An act to amend the civil practice law and rules, in relation to time limitations for filing claims for certain injuries PURPOSE OR GENERAL IDEA OF BILL: To amend the Civil Practice Law and Rules in relation to time limita- tions for filing claims for certain injuries. SUMMARY OF SPECIFIC PROVISIONS: Subdivision 4 of section 214-c of the Civil Practice Laws and Rules, as added by chapter 682 of the laws of 1986 is amended.