Bill S1249 version original - James Tedisco
BILL NUMBER: S1249 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the agriculture and markets law, in relation to increas- ing penalties for the owners of dangerous dogs PURPOSE: The purpose of this bill is to deter violent offenders from purchasing and raising possibly dangerous dogs. SUMMARY OF PROVISIONS: Section 1: Amends subdivisions 6, 7 and 8 and adds subsection 15 to section 123 of the Agriculture and Markets law.