Bill S1437 version original - Joseph A. Griffo
BILL NUMBER: S1437 SPONSOR: GRIFFO TITLE OF BILL: CONCURRENT RESOLUTION OF THE SENATE AND ASSEMBLY proposing the addition of a new article 20 to the constitution, in relation to providing for recall PURPOSE: The purpose of this bill is to allow for the recall of statewide elected officials. SUMMARY OF PROVISIONS: This bill authorizes the use of recall in New York State for all state- wide elected officers. Recall authorizes electors to remove an elective officer.