Bill S1293 version original - Kevin S. Parker
BILL NUMBER: S1293 SPONSOR: PARKER TITLE OF BILL: An act directing the department of public service to prepare a written report on the affordability of utility services PURPOSE OF BILL: The purpose of the legislation is to conduct a comprehensive study of the effect of the COVID-19 state of emergency, on utilities and consum- ers. SUMMARY OF SPECIFIC PROVISIONS: Section 1 provides the legislative findings. Section 2 lays out the definitions applicable to this act.