Bill S1285 version original - Michelle Hinchey
BILL NUMBER: S1285 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the education law, in relation to the component school districts' share of the capital expenditures of a board of cooperative educational services PURPOSE: The purpose of this legislation is to exempt certain BOCES capital expenditures from limitations upon school district tax levies. SUMMARY OF PROVISIONS: Section 1.