Bill S1340 version original - Samra Brouk
BILL NUMBER: S1340 SPONSOR: BROUK TITLE OF BILL: An act to amend the environmental conservation law, in relation to the sale of cosmetic products containing trace amounts of mercury PURPOSE OR GENERAL IDEA OF BILL: This is a chapter amendment that makes changes to provisions of chapter 754 of the laws of .2022 to make technical corrections and align the provisions with existing federal regulations. SUMMARY OF PROVISIONS: Section 1 makes technical corrections.