Bill S1314 version original - Robert Ortt
BILL NUMBER: S1314 SPONSOR: ORTT TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to fees charged for license plates PURPOSE: To amend the vehicle and traffic regarding fees charged for license plates. SUMMARY OF PROVISIONS: Section 1 requires the DMV commissioner to waive the $25 dollar fee for a new license plate for residents forced to get new license plates in the event that the state mandates a new plate design. Section 2 sets the effective date.