Bill S1463 version original - José M. Serrano
BILL NUMBER: S1463 SPONSOR: SERRANO TITLE OF BILL: An act to amend the public authorities law, in relation to financing the design, acquisition, construction, improvement and installation of East River Esplanade PURPOSE OR GENERAL IDEA OF BILL: To ensure that the funding needs for the East River Esplanade are iden- tified. SUMMARY OF SPECIFIC PROVISIONS: This bill would amended § 1285-v to the Public Authorities Law.