Bill S1526 version original - Robert Ortt
BILL NUMBER: S1526 SPONSOR: ORTT TITLE OF BILL: An act to amend the tax law, in relation to a deduction for small busi- ness employers for certain wages and compensation paid to employees PURPOSE: Allows for employers to deduct an additional amount of twenty five percent of employees wages. SUMMARY OF PROVISIONS: Section 1 amends section 612 of the tax law to allow for small business to deduct up to ten employees who make up to one hundred and ten percent of that tax year's.