Bill S1557 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1557 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the racing, pari-mutuel wagering and breeding law and the tax law, in relation to prohibiting individuals under the age of twenty-one from gambling PURPOSE: To prohibit individuals under the age of 21 from gambling. SUMMARY OF PROVISIONS: Amends Subdivision 2 of section 108 of the racing, pari-mutuel wagering: and breeding Low, Is amended to include individuals between 18 and 21 years of age from engaging in gambling.