Bill S1556 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1556 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the penal law, in relation to on duty auxiliary police officers PURPOSE: To create penal law offenses for criminal conduct directed at on-duty auxiliary police officers. SUMMARY OF PROVISIONS: Section 1 of this bill amends section 10.00 of the Penal Law to add a new subdivision 22 to provide the definition for "on-duty auxiliary police officer." Section 2 of this bill amends paragraphs (b) and (c) of subdivision 1 of section 70.