Bill S1475 version original - Roxanne J. Persaud
BILL NUMBER: S1475 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the insurance law, in relation to the mandatory coverage of hearing aids by insurers and other organizations PURPOSE: To require all managed care organizations that provide hospital, medical, or surgical benefits to individuals or groups under contracts that are issued in this state to provide coverage for prescription hear- ing aids for children. SUMMARY OF SPECIFIC PROVISIONS: Section 1.