Bill S1494 version original - Leroy Comrie
BILL NUMBER: S1494 SPONSOR: COMRIE TITLE OF BILL: CONCURRENT RESOLUTION OF THE SENATE AND ASSEMBLY proposing an amendment to section 4 of article 5 of the constitution, in relation to election of public service commissioners PURPOSE: This amendment provides for the election of the Public Service Commis- sioners as the Legislature shall so prescribe by law. Under the existing section 4 of Article 5 of the Constitution, the Public Service commis- sioners are appointed by the Governor.