Bill S1580 version original - Robert Ortt
BILL NUMBER: S1580 SPONSOR: ORTT TITLE OF BILL: An act to amend chapter 363 of the laws of 1982 relating to incorpo- ration of the Volunteer and Exempt Firemen's Benevolent Association of Lewiston Fire Company No. 2, Inc., in relation to its purpose and the use of foreign fire insurance premium taxes PURPOSE: To authorize the Lewiston Volunteer Firefighters" Benevolent Association No.2 to provide the relief, aid and assistance of members and their families who are injured.