Bill S1565 version original - James Tedisco
BILL NUMBER: S1565 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the agriculture and markets law and the administrative code of the city of New York, in relation to requiring pet dealers and pet shops to provide certain breeder and health information to purchas- ers of animals PURPOSE: To enact "Byron's Law" which would provide greater disclosure of breed- ing information to the consumer who purchases a pet at a retail pet store. SUMMARY OF PROVISIONS: § 1.