Bill S1560 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1560 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to restrictions and penalties for certain transactions involving used catalytic convert- ers PURPOSE: To help prevent the thefts of catalytic converters. SUMMARY OF PROVISIONS: Defines who can remove, sell, and/or purchase catalytic converters and/or the nonferrous metal components from a catalytic converter. Requires a permit or title of ownership to sell the converter/metals.