Bill S1473 version original - Neil D. Breslin
BILL NUMBER: S1473 SPONSOR: BRESLIN PURPOSE: Requires that instruction in the laws relating to handicapped parking be included in the pre-licensing driver's course. SUMMARY OF PROVISIONS: Amends paragraph (b) of subdivision 4 of section 502 of the vehicle and traffic law, as amended by chapter 355 of the laws of 2017, by adding a new paragraph (i). JUSTIFICATION: Violations of the handicapped parking laws are frequent, inconsiderate, and unnecessary.