Bill S1542 version original - Leroy Comrie
BILL NUMBER: S1542 SPONSOR: COMRIE TITLE OF BILL: An act to amend the public authorities law, in relation to requiring the metropolitan commuter transportation authority to obtain approval from the state comptroller to increase fares PURPOSE: Requires any proposed increase in fares by the Metropolitan Commuter Transportation Authority to be submitted to the comptroller for approval. SUMMARY OF PROVISIONS: Section 1.