Bill S1559 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1559 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the election law, in relation to the content of state- ments of campaign receipts, contributions, transfers and expenditures to and by political committees PURPOSE: This bill is designed to increase the disclosure of contributor employ- ment and campaign expenditure information. SUMMARY OF PROVISIONS: Section 1.