Bill S1563 version original - James Tedisco
BILL NUMBER: S1563 SPONSOR: TEDISCO TITLE OF BILL: An act making an appropriation for the purpose of establishing and expanding training programs for direct support professionals at boards of cooperative educational services PURPOSE: Relates to making an appropriation for the purpose of establishing and expanding training programs for direct support professionals at boards of cooperative educational services. SUMMARY OF PROVISIONS: § 1.