Bill S1495 version original - Leroy Comrie
BILL NUMBER: S1495 SPONSOR: COMRIE TITLE OF BILL: An act to amend the insurance law, in relation to enacting the small insurer protection act PURPOSE: To protect and encourage the creation of small insurance companies that write property and casualty insurance that will benefit the public and policyholders through increased competition. SUMMARY OF PROVISIONS: Section one 'of the bill states that the act shall be known and may be cited as the "small insurer protection act.