Bill S1446 version original - James Sanders Jr.
BILL NUMBER: S1446 SPONSOR: SANDERS TITLE OF BILL: An act to amend the tax law and the state finance law, in relation to gifts to the affordable housing for veterans and seniors fund PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to establish a fund for the creation of affordable housing. SUMMARY OF SPECIFIC PROVISIONS: The tax law is amended by adding a new section 630-f and the state finance law is also amended by, adding a new section 99-bb.