Bill S1561 version original - James Tedisco
BILL NUMBER: S1561 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the agriculture and markets law, in relation to aggra- vated neglect of a companion animal PURPOSE OR GENERAL IDEA OF BILL: Creates the crime of Aggravated Neglect of a Companion Animal. SUMMARY OF PROVISIONS: Section 1 creates new law, Section 355-a of the Agriculture and Markets law. This section creates the crime of Aggravated Neglect of a Companion Animal.