Bill S1454 version original - José M. Serrano
BILL NUMBER: S1454 SPONSOR: SERRANO TITLE OF BILL: An act to amend the penal law, in relation to establishing the crime of disruption of an online public meeting PURPOSE OR GENERAL IDEA OF BILL: To increase penalties associated with disrupting a public meeting when the intent is to cause public inconvenience, annoyance or alarm, without lawful authority, acting through a commuter service. JUSTIFICATION: The current public health crisis has increased the use of online video communications.