Bill S1470 version original - Neil D. Breslin
BILL NUMBER: S1470 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to physical therapy services PURPOSE: To limit the imposition of co-payments for physical therapy services to no more than twenty percent of the reimbursement to the provider of care. SUMMARY OF PROVISIONS: Sections 1 through 5 prohibit payors from imposing costs on insureds for the provision of physical therapy services in excess of 20 percent of the reimbursement to the provider of care.