Bill S1603 version original - Joseph A. Griffo
BILL NUMBER: S1603 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the penal law, in relation to establishing the crime of making a threat of mass violence PURPOSE: The purpose of this legislation is to criminalize threats of mass violence against schools, colleges and universities, places of worship, businesses and mass gatherings. SUMMARY OF PROVISIONS: The bill establishes a new crime of making a threat of mass violence.