Bill S1553 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1553 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the education law, in relation to requiring that public and nonpublic schools be locked during the school day PURPOSE: To prevent unwanted or potentially dangerous intruders from entering schools while classes are in session. SUMMARY OF PROVISIONS: Section 1- Amends section 305 of the education law by adding a new subdivisionæ57 to require that entrance doors to all schools be kept locked during the school day.