Bill S1642 version original - James Tedisco
BILL NUMBER: S1642 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the executive law, the general municipal law and the town law, in relation to requiring police officers to attend annual diversity training seminars PURPOSE OR GENERAL IDEA OF BILL: To require police officers to attend annual diversity training seminars SUMMARY OF PROVISIONS: § 1.This act shall be known and may be cited as the "law enforcement diversity and conflict mitigation training act". § 2.