Bill S1711 version original - Leroy Comrie
BILL NUMBER: S1711 SPONSOR: COMRIE TITLE OF BILL: An act to amend the not-for-profit corporation law and the religious corporations law, in relation to the correction of tax errors made by the nonprofit revitalization act SUMMARY OF MAJOR PROVISIONS: SECTION ONE WOULD CORRECT AN OBVIOUS ERROR IN NOT-FOR-PROFIT CORPORATION LAW SECTION 906(A) (1)(D) TO CHANGE THE CROSS REFERENCE FROM PARAGRAPH (C) OF SECTION 607 TO PARAGRAPH (B). THERE IS NO PARAGRAPH (C).