Bill S1670 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1670 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the election law, in relation to ballots that contain the name of a candidate who previously withdrew from an election PURPOSE: To disqualify a vote cast for a candidate who has voluntarily withdrawn from consideration prior to the election. SUMMARY OF PROVISIONS: Section 1.