Bill S1620 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1620 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the penal law, in relation to including certain activity in transit and transportation facilities within the crime of criminal trespass in the second degree PURPOSE: To establish a criminal penalty for trespassing and vandalism of a public transportation system which results in damage and disruption of service. SUMMARY OF PROVISIONS: Section 1 - Subdivision 2 of section 140.