Bill S1706 version original - Leroy Comrie
BILL NUMBER: S1706 SPONSOR: COMRIE TITLE OF BILL: An act to amend the mental hygiene law, in relation to reimbursement for certain costs and expenses related to the novel coronavirus, COVID-19 PURPOSE: Authorizes voluntary providers of supports and services to individuals with intellectual and developmental disabilities to bill and be reim- bursed within thirty days of billing for novel coronavirus, COVID-19 related costs and expenses. SUMMARY OF PROVISIONS: Section 1.