Bill S1725A version original - Pete Harckham
BILL NUMBER: S1725 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the environmental conservation law, in relation to the protection of certain streams PURPOSE: This act will reclassify class C waterways as streams for the purpose of protection. SUMMARY OF PROVISIONS: Section 1 adds waterways classified as "C" to the definition of streams.