Bill S1672 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1672 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the highway law, in relation to providing notice to school districts and nonpublic schools located in a city with a popu- lation of one million or more of certain construction projects PURPOSE: To require the written notification to a school district 14 days prior to the commencement of a construction project that may disrupt the traf- fic to and from a school. SUMMARY OF PROVISIONS: Section 1.