Bill S1769 version original - Cordell Cleare
BILL NUMBER: S1769 SPONSOR: CLEARE TITLE OF BILL: An act to amend the general business law, in relation to requiring tele- marketers to add a customer's telephone number to all do-not-call lists maintained by such telemarketer PURPOSE: To allow an individual to sign up for more than one do not call list at a time. SUMMARY OF PROVISIONS: Amends the BGL to allow an individual to sign up for more than one do not call list at a time.