Bill S1660 version original - Jamaal Bailey
BILL NUMBER: S1660 SPONSOR: BAILEY TITLE OF BILL: An act to amend the criminal procedure law, in relation to the effect of termination of criminal actions and proceedings against a minor PURPOSE OR GENERAL IDEA OF BILL: Allows for minors with no prior arrest or criminal records, who are found innocent or have a case against them dismissed, to have their arrest record, fingerprints, and photographs expunged. SUMMARY OF PROVISIONS: Section 160.