Bill S1676 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1676 SPONSOR: ADDABBO TITLE OF BILL: An act establishing a temporary task force on variable supplements fund PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to establish a task force to make a final determination of whether or not disabled NYPD and FDNY retirees are entitled to receive VSF supplementation benefits. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Legislative intent.