Bill S1657 version original - Robert Ortt
BILL NUMBER: S1657 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to qualified firearms instructors PURPOSE: To authorize pre-license possession of a pistol or revolver at an indoor or outdoor pistol range under the supervision of a qualified firearms instructor. SUMMARY OF PROVISIONS: Section 1 amends subdivision 19 of section 265.00 of the Penal Law to add qualified firearms instructor to the definition of "duly authorized instructor".