Bill S1703 version original - Robert Ortt
BILL NUMBER: S1703 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to licenses to carry and possess firearms PURPOSE: This legislation allows for the holder of a pistol license to carry the weapon concealed. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill removes language that limits the individuals who are eligible to carry a pistol concealed.