Bill S1619 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1619 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the environmental conservation law, in relation to prohibiting the use of leg-gripping traps PURPOSE: To prohibit the use of wildlife leg-gripping traps. SUMMARY OF SPECIFIC PROVISIONS: Section one amends subdivisions 5 and 6 of 11-1101 of the environmental conservation law. Section two amends subdivision 7 of section 11-1901 of the environmental conservation law.