Bill S1775 version original - Jamaal Bailey
BILL NUMBER: S1775 SPONSOR: BAILEY TITLE OF BILL: An act to amend the multiple dwelling law, in relation to landlord checks of the state sex offender registry and notification of certain tenants PURPOSE OR GENERAL IDEA OF BILL: Requires landlords to check state offender registry when leasing to a new tenant or on notice of sublet, and to notify tenants with children of any listing. SUMMARY OF SPECIFIC PROVISIONS: Section 1.