Bill S1661 version original - Jamaal Bailey
BILL NUMBER: S1661 SPONSOR: BAILEY TITLE OF BILL: An act to amend the tax law, in relation to providing for a deduction from personal gross income for expenses incurred in the adoption of a child PURPOSE: To allow those who adopt a child to deduct from their gross income tax, any and all costs incurred in legal fees, agency costs, etc. in the adoption proceedings. JUSTIFICATION: Expenses incurred for adoption services should not be taxable.