Bill S1627 version original - Leroy Comrie
BILL NUMBER: S1627 SPONSOR: COMRIE TITLE OF BILL: An act to amend the insurance law, in relation to certain policy review PURPOSE OR GENERAL IDEA OF BILL: To provide auto insurance policy holders in the assigned risk plan with increased access to auto coverage in the voluntary market. SUMMARY OF PROVISIONS: Section one of the bill would add a new section 5305 to the Insurance Law.