Bill S1699 version original - Pamela Helming
BILL NUMBER: S1699 SPONSOR: HELMING TITLE OF BILL: An act to amend the public health law, in relation to reporting inci- dents of possible professional misconduct PURPOSE: To enact Clara's Law requiring hospitals and healthcare facilities to report incidents of a sexual offense to the State Departments of Health and Education. SUMMARY OF PROVISIONS: Section 1: Establishes the title "Clara's Law.