Bill S1636 version original - Peter Oberacker
BILL NUMBER: S1636 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the public health law, in relation to prohibiting a mandatory immunization against the coronavirus PURPOSE: The general idea of this bill is to allow the Coronavirus vaccine to remain taken only on a voluntary basis. SUMMARY OF PROVISIONS: Section 1. Section 2182 of the public health law is renumbered section 2184 and a new section 2182 is added Section 2.