Bill S1644 version original - Brad Hoylman-Sigal
BILL NUMBER: S1644 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the penal law, in relation to creating the offense of fostering the sale of stolen goods PURPOSE OR GENERAL IDEA OF BILL: To prevent the use of online marketplaces and other forums for the fenc- ing of stolen goods. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill amends the penal law to create the offense of fostering the sale of stolen goods, which will be a class A misdemeanor.