Bill S1780A version original - Jamaal Bailey
BILL NUMBER: S1780 SPONSOR: BAILEY TITLE OF BILL: An act to incentivize the purchase of local produce in supermarkets and grocery stores through the FreshConnect program PURPOSE OR GENERAL IDEA OF BILL: To expand the reach of the FreshConnect program across the state and increase access to affordable fruits and vegetables to New Yorkers in need. SUMMARY OF SPECIFIC PROVISIONS: Section 1 establishes definitions for local produce and producer.