Bill S1746 version original - Jessica Ramos
BILL NUMBER: S1746 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to developing and implement- ing programs to prevent workplace violence in public schools PURPOSE: This bill amends section 27-b of the Labor Law to include public schools within the provisions of current law designed to require public employ- ers to develop and implement programs to prevent workplace violence. SUMMARY OF PROVISIONS: An amendment is made within section 27-b of the Labor law.