Bill S1771 version original - Cordell Cleare
BILL NUMBER: S1771 SPONSOR: CLEARE TITLE OF BILL: An act to amend the correction law, in relation to commencing a study on the relationship between the corrections system and the child welfare system PURPOSE OR GENERAL IDEA OF BILL: To conduct a study on the relationship between the corrections systems and the child welfare system by enabling data collection on inmates. SUMMARY OF SPECIFIC PROVISIONS: Section 1: The correction law is amended by adding a new section 30 as follows.