Bill S1760 version original - James Sanders Jr.
BILL NUMBER: S1760 SPONSOR: SANDERS TITLE OF BILL: An act to amend the executive law, the civil rights law and the educa- tion law, in relation to certain discriminatory practices; and to amend the education law, in relation to the hiring and qualification of professionals in education PURPOSE: This legislation expands discrimination protections under New York State law and provides for greater transparency in the hiring process of education professionals. SUMMARY: Section 1. Short title.