Bill S1714 version original - Monica Martinez
BILL NUMBER: S1714REVISED 1/20/23 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend chapter 397 of the laws of 1996, relating to authorizing the town of Islip to lease certain lands to a business corporation, in relation to the lease term PURPOSE: The purpose of this legislation is to grant authority to the Town of Islip to extend an existing lease for The Rinx in Hauppauge. SUMMARY OF PROVISIONS: Section 1.