Bill S1940 version original - Jamaal Bailey
BILL NUMBER: S1940 SPONSOR: BAILEY TITLE OF BILL: An act to amend the tax law, in relation to taxpayers affected by a federal government shutdown PURPOSE OR GENERAL IDEA OF BILL: To allow New York State residents affected by a federal government shut- down an automatic tax return extension that otherwise is required to file a return on April fifteenth. The length of the extension is sixty days from the date the federal government reopens with no penalties or interested imposed on the taxpayer.