Bill S1871 version original - Cordell Cleare
BILL NUMBER: S1871 SPONSOR: CLEARE TITLE OF BILL: An act to amend the tax law and the state finance law, in relation to establishing the public housing fund PURPOSE OR GENERAL IDEA OF BILL: This bill establishes a Public Housing Fund, to help save and support Public Housing. The source of funding will be a 10% infusion from the Real Estate Transfer Tax. SUMMARY OF PROVISIONS: Directs 10% of the Real Estate Transfer Tax into a Public Housing Fund.