Bill S1919 version original - Peter Oberacker
BILL NUMBER: S1919 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the regis- tration of vintage snowmobiles PURPOSE: To exempt those sleds manufactured prior to the last thirty years. SUMMARY OF PROVISIONS: Amends subdivision 11 of section 2222 of the vehicle and traffic law by adding a new paragraph (d). JUSTIFICATION: The collecting and restoration of vintage sleds, has become a popular pastime in the snowmobile community.