Bill S1897 version original - Toby Ann Stavisky
BILL NUMBER: S1897 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the real property tax law, in relation to limitations on assessed value for any parcel that is held in cooperative or condominium forms of ownership PURPOSE OR GENERAL IDEA OF BILL: Relates to limitations on assessed value for any parcel that is held in cooperative or condominium forms of ownership. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Section 1805 (3) of the Property Tax Law by adding a new subdivision 7.