Bill S1913 version original - Robert Ortt
BILL NUMBER: S1913 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to aggravated harassment of an employee by an incarcerated individual PURPOSE: This legislation removes language as it relates to causing an employee to come into contact with certain body fluid by an inmate. SUMMARY OF SPECIFIC PROVISIONS: Section 1 Amends Section 240.