Bill S1807 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1807 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the tax law, in relation to requiring the state to keep state lottery winners identities anonymous to the general public unless the lottery winner gives their consent PURPOSE OR GENERAL IDEA OF BILL: To shield the identity of lottery winners for their own personal privacy and safety. SUMMARY OF PROVISIONS: To prevent the Division of the Lottery from publicly disclosing the name, address, or other identifying.