Bill S1832 version original - Joseph A. Griffo
BILL NUMBER: S1832 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to distinctive plates for legion of valor recipients PURPOSE: Authorizes the issuance of distinctive legion of valor license plates. SUMMARY OF PROVISIONS: Amends the vehicle and traffic law to establish the "Legion of Valor" license plate for Medal of Honor and Service Cross recipients and provides that no registration and/or service fee will be charged for the issuance of such plates.