Bill S1970 version original - Luis R. Sepúlveda
BILL NUMBER: S1970 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the correction law, in relation to failure to complete a program for time allowances due to circumstances beyond an individual's control PURPOSE: To waive the requirement that incarcerated individuals complete certain programs if such failure to complete is due to circumstances beyond their control. SUMMARY OF PROVISIONS: Sections 1 and 2 amend Correction Law § 803(1)(a). Sections 3 and 4 amend Correction Law § 803(1)(d)(iv).