Bill S1962 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1962 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the racing, pari-mutuel wagering and breeding law, in relation to the definition of wager and the licensing of mobile sports wagering operators PURPOSE: To update the law related to promotional credits, to codify the existing tax rates, and to license additional operators. SUMMARY OF PROVISIONS: Section 1: Adds "thing of value" to the definition of sports wager.