Bill S1842 version original - Leroy Comrie
BILL NUMBER: S1842 SPONSOR: COMRIE TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to requiring motor carriers of passengers to post, for public viewing, certain infor- mation on their websites PURPOSE OR GENERAL IDEA OF BILL: To require bus carriers to post the safety records of their bus company on the company's website, as listed on the Federal Motor Carrier Safety Administration's website. SUMMARY OF SPECIFIC PROVISIONS: Section 1.