Bill S1828 version original - Mario Mattera
BILL NUMBER: S1828 SPONSOR: MATTERA TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to requiring that driver's licenses indicate the blood type of the licensee upon the licensee's request PURPOSE: To allow a driver to indicate their blood type on their driver's license if they choice to. SUMMARY OF PROVISIONS: Section 1 amends by adding a new paragraph to subdivision 1 of section 504 of the vehicle and traffic law.