Bill S1869 version original - Cordell Cleare
BILL NUMBER: S1869 SPONSOR: CLEARE TITLE OF BILL: An act to amend the New York city charter, in relation to requiring the commissioner of the department of housing preservation and development to issue an annual report on the tenant interim lease program PURPOSE: To require the commissioner of the department of housing preservation and development to issue an annual report on the tenant interim lease program. SUMMARY OF PROVISIONS: To require annual reporting on the tenant interim lease program.