Bill S1966 version original - Luis R. Sepúlveda
BILL NUMBER: S1966REVISED 11/9/23 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the penal law, in relation to individuals engaged in prostitution who are victims of or witnesses to a crime PURPOSE OR GENERAL IDEA OF BILL: To encourage people in prostitution who are crime victims and witnesses to seek help without fear of being prosecuted for prostitution. SUMMARY OF PROVISIONS: Section 1 creates a new Penal Law § 230.45 to provide immunity from prosecution under § 230.