Bill S1892 version original - Toby Ann Stavisky
BILL NUMBER: S1892 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the penal law, in relation to adding ammunition to the list of prohibited items to purchase or dispose of after certain convictions PURPOSE OR GENERAL IDEA OF BILL: To prevent "straw purchases" of ammunition. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends Section 265.17 of the penal law, as amended by chapter 1 of the laws of 2013, by adding ammunition to the crime of criminal purchase or disposal of a weapon.