Bill S1836 version original - Joseph A. Griffo
BILL NUMBER: S1836 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the highway law and the vehicle and traffic law, in relation to the maintenance, improvement and repair of certain town highways PURPOSE: This bill authorizes towns to designate certain town roads as low-volume roads and certain low-volume roads as minimum maintenance roads. SUMMARY OF PROVISIONS: Section 1 would add a new section 205-C to the highway law.