Bill S1845 version original - Leroy Comrie
BILL NUMBER: S1845 SPONSOR: COMRIE TITLE OF BILL: An act to amend the tax law, in relation to a tax on gross income upon every corporation which derives income from the data individuals of this state share with such corporations; and to amend the state finance law, in relation to establishing the New York data fund SUMMARY OF SPECIFIC PROVISIONS: Section 1. Section 209 of the tax law is amended by adding a new subdi- vision 13 to read as follows: 13.