Bill S1908 version original - Cordell Cleare
BILL NUMBER: S1908 SPONSOR: CLEARE TITLE OF BILL: An act to amend the civil service law, in relation to reporting the race, ethnicity and salary ranges for women employed in service of the state PURPOSE OR GENERAL IDEA OF BILL: Requires the NYS Department of Civil Service to include ethnicity and race data in its annual report of women in state government workforce. The data is currently omitted. SUMMARY OF SPECIFIC PROVISIONS: Section 1.