Bill S1965A version original - Joseph P. Addabbo Jr.
BILL NUMBER: S1965 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the insurance law, in relation to coverage for prenatal vitamins SUMMARY OF PROVISIONS: Section 1. Describes which individual medical policies must provide coverage for prenatal vitamins. Coverage of the prenatal vitamins are subject to annual deductibles and coinsurance that are consistent with those established for a given policy.